seasonal-deceptive-website-detector/requirements.txt
# Streamlit Framework
streamlit==1.28.1

# Web Scraping & Analysis
requests==2.31.0
beautifulsoup4==4.12.2

# Domain & WHOIS Information
python-whois==0.9.3

# Database
mysql-connector-python==8.2.0

# Data Processing
pandas==2.1.1
numpy==1.24.3

# SSL/TLS Certificate Analysis
cryptography==41.0.5

# Utilities
python-dateutil==2.8.2